+++ skimage/filters/rank/percentile_cy.pyx
@@ -9,7 +9,7 @@ from .core_cy cimport dtype_t, dtype_t_out, _core
 
 cdef inline void _kernel_autolevel(dtype_t_out* out, Py_ssize_t odepth,
                                    Py_ssize_t* histo,
-                                   double pop, dtype_t g,
+                                   Py_ssize_t pop, dtype_t g,
                                    Py_ssize_t max_bin, Py_ssize_t mid_bin,
                                    double p0, double p1,
                                    Py_ssize_t s0, Py_ssize_t s1) nogil:
@@ -43,7 +43,7 @@ cdef inline void _kernel_autolevel(dtype_t_out* out, Py_ssize_t odepth,
 
 cdef inline void _kernel_gradient(dtype_t_out* out, Py_ssize_t odepth,
                                   Py_ssize_t* histo,
-                                  double pop, dtype_t g,
+                                  Py_ssize_t pop, dtype_t g,
                                   Py_ssize_t max_bin, Py_ssize_t mid_bin,
                                   double p0, double p1,
                                   Py_ssize_t s0, Py_ssize_t s1) nogil:
@@ -72,7 +72,7 @@ cdef inline void _kernel_gradient(dtype_t_out* out, Py_ssize_t odepth,
 
 cdef inline void _kernel_mean(dtype_t_out* out, Py_ssize_t odepth,
                               Py_ssize_t* histo,
-                              double pop, dtype_t g,
+                              Py_ssize_t pop, dtype_t g,
                               Py_ssize_t max_bin, Py_ssize_t mid_bin,
                               double p0, double p1,
                               Py_ssize_t s0, Py_ssize_t s1) nogil:
@@ -98,7 +98,7 @@ cdef inline void _kernel_mean(dtype_t_out* out, Py_ssize_t odepth,
 
 cdef inline void _kernel_sum(dtype_t_out* out, Py_ssize_t odepth,
                              Py_ssize_t* histo,
-                             double pop, dtype_t g,
+                             Py_ssize_t pop, dtype_t g,
                              Py_ssize_t max_bin, Py_ssize_t mid_bin,
                              double p0, double p1,
                              Py_ssize_t s0, Py_ssize_t s1) nogil:
@@ -124,7 +124,7 @@ cdef inline void _kernel_sum(dtype_t_out* out, Py_ssize_t odepth,
 
 cdef inline void _kernel_subtract_mean(dtype_t_out* out, Py_ssize_t odepth,
                                        Py_ssize_t* histo,
-                                       double pop, dtype_t g,
+                                       Py_ssize_t pop, dtype_t g,
                                        Py_ssize_t max_bin,
                                        Py_ssize_t mid_bin, double p0,
                                        double p1, Py_ssize_t s0,
@@ -151,7 +151,7 @@ cdef inline void _kernel_subtract_mean(dtype_t_out* out, Py_ssize_t odepth,
 
 cdef inline void _kernel_enhance_contrast(dtype_t_out* out,
                                           Py_ssize_t odepth,
-                                          Py_ssize_t* histo, double pop,
+                                          Py_ssize_t* histo, Py_ssize_t pop,
                                           dtype_t g,
                                           Py_ssize_t max_bin,
                                           Py_ssize_t mid_bin, double p0,
@@ -188,7 +188,7 @@ cdef inline void _kernel_enhance_contrast(dtype_t_out* out,
 
 cdef inline void _kernel_percentile(dtype_t_out* out, Py_ssize_t odepth,
                                     Py_ssize_t* histo,
-                                    double pop, dtype_t g,
+                                    Py_ssize_t pop, dtype_t g,
                                     Py_ssize_t max_bin, Py_ssize_t mid_bin,
                                     double p0, double p1,
                                     Py_ssize_t s0, Py_ssize_t s1) nogil:
@@ -213,7 +213,7 @@ cdef inline void _kernel_percentile(dtype_t_out* out, Py_ssize_t odepth,
 
 cdef inline void _kernel_pop(dtype_t_out* out, Py_ssize_t odepth,
                              Py_ssize_t* histo,
-                             double pop, dtype_t g,
+                             Py_ssize_t pop, dtype_t g,
                              Py_ssize_t max_bin, Py_ssize_t mid_bin,
                              double p0, double p1,
                              Py_ssize_t s0, Py_ssize_t s1) nogil:
@@ -234,7 +234,7 @@ cdef inline void _kernel_pop(dtype_t_out* out, Py_ssize_t odepth,
 
 cdef inline void _kernel_threshold(dtype_t_out* out, Py_ssize_t odepth,
                                    Py_ssize_t* histo,
-                                   double pop, dtype_t g,
+                                   Py_ssize_t pop, dtype_t g,
                                    Py_ssize_t max_bin, Py_ssize_t mid_bin,
                                    double p0, double p1,
                                    Py_ssize_t s0, Py_ssize_t s1) nogil:
